requirements.txt
streamlit
langchain
openai
faiss-cpu
python-dotenv
langchain-community
unstructured
tiktoken
transformers
torch
spacy